Allow navigation up & down within element groups.
So from neon, you can easily navigate up to helium or down to argon.

Although the lanthanides and actinides do not strictly belong to any groups,
this change allows you to navigate up and down between them.

The goal here is to make it easy to navigate between elements that behave
similarly due to their electron configurations.