Provide better feedback when bumping up against search constraints.
Searching for a single character that isn't an element symbol doesn't show any
results due to my design choices, not because no element names, molecule names,
or formulas contain that character.

In my opinion, if someone searches for a letter that happens to be an element
symbol, they probably want to see the element, and all the formulas that
contain that element, but not every molecule name that contain that letter.